requirements.txt
langchain
google-generativeai
github3.py
gitpython
python-dotenv
